tests/wpt/tests/css/css-sizing/stretch/indefinite-5.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/css-sizing-4/#stretch-fit-sizing">
<script src='/resources/testharness.js'></script>
<script src='/resources/testharnessreport.js'></script>
<script src="/resources/check-layout-th.js"></script>
<meta name="assert" content="height:stretch inside a flex item behaves as auto when the flex container cross size is indefinite">

<body onload="checkLayout('[data-expected-client-height]')">

  <div style="display: flex;">
    <div>
      <div style="display: flex; height: stretch;" data-expected-client-height="0"></div>
    </div>
  </div>

  <div style="display: flex;">
    <div>
      <div style="height: stretch;" data-expected-client-height="20">
        <div style="height: 20px;"></div>
      </div>
    </div>
  </div>
